Fixes #16

+++ GameSaveManager.Core/Utils/FileSystemUtils.cs
@@ -5,15 +5,40 @@ namespace GameSaveManager.Core.Utils
 {
     public static class FileSystemUtils
     {
-        public static string GetGameFolderLocationAppData(string folderName) => Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), folderName);
-
-        public static bool CheckFileExistence(string path) => File.Exists(path);
-
-        public static bool DeleteZipFile(string path)
+        public static string FindPath(string folder)
         {
-            File.Delete(path);
+            return Directory.Exists(Path.Combine(GetGameFolderLocationAppData(), folder))
+                ? Path.Combine(GetGameFolderLocationAppData(), folder)
+                : Directory.Exists(Path.Combine(GetDocumentsFolder(), folder))
+                    ? Path.Combine(GetDocumentsFolder(), folder)
+                    : null;
+        }
 
-            return CheckFileExistence(path);
+        public static string GetTempFolder()
+        {
+            return Path.GetTempPath();
+        }
+
+        public static string GetDocumentsFolder()
+        {
+            return Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+        }
+
+        public static string GetGameFolderLocationAppData()
+        {
+            return Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
+        }
+
+        public static bool CheckFileExistence(string path)
+        {
+            return File.Exists(path);
+        }
+
+        public static bool DeleteCreatedFile(string path)
+        {
+            if (CheckFileExistence(path)) File.Delete(path);
+
+            return !CheckFileExistence(path) || DeleteCreatedFile(path);
         }
     }
 }
